+++ src/structures/parsers.ts
@@ -1,4 +1,4 @@
-import { CustomNodesManager } from "./managers";
+import { CustomNodesManager } from './managers/index.js'
 
 export const Parsers = {
     Classes: new CustomNodesManager()